==> notebooks/02-data_cleaning.ipynb <==
# DATA CLEANING

# import modules and libraries
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns

# import dataset
data = pd.read_csv('../input/car_data.csv')
data.head(10)

The car features are names of the car, price of the car, production year, miles travelled, engine and transmission types

#get the shape of the data
data.shape

#get the statistical description of the float columns
data.describe()

data.info()

data.isnull().sum()

There are 10 missing values in the transmission column. That is a very small number so we need to check the column to see how we can clean it

data[data['transmission'].isnull()]

The rows with null values are small and look inconsistent, so it is best we delete it

data= data.dropna()
data.shape

Ten rows have been deleted which is less than 1% of the total data

#delete duplicate rows 
data = data.drop_duplicates(keep= 'first', inplace=False)

data.shape

## NAME

We will extract the name of the car manufacturers from the name column

data['manufacturer'] = data['name'].apply(lambda x: x.split(' ')[0])
data['manufacturer'].head()

data.manufacturer.nunique()

data.manufacturer.unique()

We will clean these data by making similar maufacturers uniform.

data = data.replace({'manufacturer' : 
            {'MAZDA' : 'Mazda', 'JAGUAR':'Jaguar', 'AUDI' : 'Audi', 'NISSAN': 'Nissan', 'MINI': 'Mini', 
             'VOLKSWAGEN':'Volkswagen', 'VAUXHALL':'Vauxhall', 'TOYOTA':'Toyota', 'SKODA':'Skoda', 'FORD':'Ford',
             'Bmw':'BMW','SUZUKI' : 'Suzuki', 'RENAULT':'Renault', 'PEUGEOT':'Peugeot', 'CITROEN':'Citroen',
             'VOLVO':'Volvo', 'FIAT':'Fiat', 'Ds':'DS', 'DACIA':'Dacia', 'ABARTH':'Abarth', 'SMART':'Smart', 
             'smart':'Smart','SEAT':'Seat', 'MITSUBISHI':'Mitsubishi', 'KIA':'Kia', 'HYUNDAI':'Hyundai',
             'HONDA':'Honda','MASERATI':'Maserati', 'PORSCHE':'Porsche', 'INFINITI':'Infiniti', 'Alfa':'Alfa-Romero',
             'Mercedes': 'Mercedes-Benz', 'MERCEDES-BENZ': 'Mercedes-Benz', 'Mercedes-benz': 'Mercedes-Benz',
             'Land':'Land-Rover', 'LAND': 'Land-Rover'}
             })

print(data.manufacturer.unique())
print('\n')
print('There are ' + str(data.manufacturer.nunique()) + ' different car manufacturers in the dataset')

## PRICE

data['price'] = data['price'].str.replace(',', '').str.replace('£', '').astype(np.int64)
data['price'].head()

## YEAR

We can't work with the year column like that, we have to convert it to age so that it will be easy to understand

#convert the year column to integer
data['year']=data['year'].astype(str).astype(np.int64)

Current_Year = 2021

data['age'] = Current_Year - data['year']

data['age'] = data['age'].astype(np.int64)

data['age'].head()

## MILEAGE

Remove the commas and miles from the mileage column

data['mileage'] = data['mileage'].str.replace(',', '').str.replace(' ', '').str.replace('miles', '').astype(np.int64)
data['mileage'].head()

## ENGINE

data['engine'].unique()

data = data.replace({'engine' : 
                     {'Petrol hybrid': 'Hybrid', 'Petrol hybrid': 'Hybrid', 
                      'Petrol / electric hy' : 'Hybrid', 'Petrol plug-in hybri': 'Plug_in_hybrid',
                     'Petrol/electric' : 'Hybrid'}
                     })
data['engine'].unique()

## TRANSMISSION

data['transmission'].unique()

data = data.replace({'transmission' : 
                     {'Semi auto': 'Semiautomatic', 'Semiauto': 'Semiautomatic',
                      'Semi automatic': 'Semiautomatic', 'Manual ': 'Manual',
                      'Semi-automatic': 'Semiautomatic', 'G-tronic+': 'Automatic',
                      'Cvt': 'Automatic'}
                     
                     })               

Reorder the columns to make the dataframe easy to understand

data = data[['name', 'manufacturer', 'year', 'age', 'mileage', 'engine', 'transmission', 'price']]
data.head()

data.shape

#save the cleaned data to a csv file

data.to_csv('../input/cleaned_car_data.csv', index = False)
