==> notebooks/04-model_building.ipynb <==
# MODEL BUILDING

# import the relevant libraries
import pandas as pd
import numpy as np
from sklearn.preprocessing import LabelEncoder
from sklearn.model_selection import train_test_split, GridSearchCV
from sklearn.preprocessing import MinMaxScaler
from sklearn.linear_model import LinearRegression, Ridge
from sklearn.ensemble import ExtraTreesRegressor
from lightgbm import LGBMRegressor
from xgboost import XGBRegressor
from sklearn.ensemble import RandomForestRegressor
from catboost import CatBoostRegressor
from sklearn.metrics import mean_absolute_error, accuracy_score, mean_squared_error
from sklearn import metrics
import warnings
warnings.filterwarnings('ignore')
import pickle

data = pd.read_csv('../input/cleaned_car_data.csv')
data.head()

#drop the name and year columns because it is irrelevant in our model building
data = data.drop(['name', 'year'], axis=1)

data.head()

data['manufacturer'].unique()

# label encode the categorical values
le_manufacturer = LabelEncoder()
le_engine = LabelEncoder()
le_transmission = LabelEncoder()

data['manufacturer'] = le_manufacturer.fit_transform(data['manufacturer'])
data['engine'] = le_engine.fit_transform(data['engine'])
data['transmission'] = le_transmission.fit_transform(data['transmission'])

data.head(10)

# creating X and y variables
X = data.drop('price', axis=1)

# log transform the price column
y = np.log(data['price'])

#train test split
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.3, random_state=42)

X_train.shape, X_test.shape, y_train.shape, y_test.shape

# feature scale the X_train and X_test values

norm = MinMaxScaler().fit(X_train)

# transform training data
X_train = norm.transform(X_train)

# transform testing data
X_test = norm.transform(X_test)

print(X_train)
print('\n')
print(X_test)


models = {
    'linear_model': LinearRegression(),
    'ridge_model': Ridge(random_state=123),
    'extree_model':ExtraTreesRegressor(random_state = 123),
    'lgbm_model':LGBMRegressor(random_state = 123),
    'xgboost_model':XGBRegressor(random_state = 123),
    'rf_model' : RandomForestRegressor(random_state = 123)
 }

def train_model(models: dict) -> pd.DataFrame:
    """
    It takes in a dictionary containing a key-pair of model name and estimators.
    It returns a data frame containing the metrics of the trained model.
    """
    my_dict = {}
    name_list, train_score_list, r_sqd_list, mae_list, rmse_list = [], [], [], [], []
    for name, estimator in models.items():
        # fit
        estimator.fit(X_train, y_train)

        # make predictions
        y_pred = estimator.predict(X_test)

        # metrics
        train_score = estimator.score(X_train, y_train)
        r_sqd = metrics.r2_score(y_test, y_pred)
        mae = metrics.mean_absolute_error(y_test, y_pred)
        mse = metrics.mean_squared_error(y_test, y_pred)
        rmse = np.sqrt(mse)

        # add the metrics to the empty list
        name_list.append(name)
        train_score_list.append(train_score)
        r_sqd_list.append(r_sqd)
        mae_list.append(mae)
        rmse_list.append(rmse)

    my_dict["Name"] = name_list
    my_dict["Train_Score"] = train_score_list
    my_dict["R_squared"] = r_sqd_list
    my_dict["Mean_absolute_error"] = mae_list
    my_dict["Root_mean_sqd_error"] = rmse_list

    my_dataframe = pd.DataFrame(my_dict)
    my_dataframe = my_dataframe.sort_values("Root_mean_sqd_error")
    return my_dataframe

train_model(models)

#create the grid
grid = {'max_depth': [3,4,5],'n_estimators':[100, 200, 300]}

#Instantiate GridSearchCV
model = GridSearchCV (estimator = LGBMRegressor(random_state = 123), param_grid = grid, scoring ='neg_root_mean_squared_error', cv = 5)

model.fit(X_train,y_train, verbose = False)

y_pred = model.predict(X_test)

grid_model = pd.DataFrame({
    'model': ['LGBM'],
    'r_squared': [metrics.r2_score(y_test, y_pred)],
    'mae': [mean_absolute_error(y_test, y_pred)],
    'rmse': [np.sqrt(metrics.mean_squared_error(y_test, y_pred))]
    })
grid_model


#### Make Predictions on new data.

# manufacturer, age, mileage, engine, transmission
new_data = [['Mazda', 14, 63131, 'Petrol', 'Manual']]
# convert to array
new_data = np.array(new_data)
new_data

new_data[:,0] = le_manufacturer.transform(new_data[:,0])
new_data[:,3] = le_engine.transform(new_data[:,3])
new_data[:,4] = le_transmission.transform(new_data[:,4])
new_data

new_data = norm.transform(new_data)
new_data

# make predictions
price = model.predict(new_data)
# convert the price from log_price to actual price
act_price = np.exp(price) + 1
act_price = round(act_price[0])  # round to the neaarest pounds

print(f"The estimated cost of the car is {act_price:,} pounds")

data = {"model": model, "normalization": norm}
with open('../models/regressor.pkl', 'wb') as file:
    pickle.dump(data, file)
